+++ src/Index.ts
@@ -16,4 +16,5 @@ export * from './TinodeLog.ts';
 export * from './TinodeMessage.ts';
 export * from './TinodeStorage.ts';
 export * from './TinodeTopics.ts';
+export * from './ImAttachment.ts';
 export * from './Tinode.ts';